api_test_v3/requirements.txt
requests>=2.25.0
mysql-connector-python>=8.0.0